+++ lib/src/command/command.dart
@@ -5,6 +5,7 @@ import 'package:ff_annotation_route/src/command/route_helper.dart';
 import 'package:ff_annotation_route/src/command/route_names.dart';
 import 'package:ff_annotation_route/src/command/save.dart';
 import 'package:ff_annotation_route/src/command/settings_no_arguments.dart';
+import 'package:io/ansi.dart';
 
 abstract class Command {
   /// full command
@@ -28,7 +29,7 @@ abstract class Command {
   String get command => '$short${' ' * (3 - short.length)}, $full';
 }
 
-final List<Command> commands = [
+final List<Command> ffCommands = [
   Help(),
   Path(),
   RouteConstants(),
@@ -42,7 +43,7 @@ List<Command> initCommands(List<String> arguments) {
   final result = <Command>[];
   for (var i = 0; i < arguments.length; i++) {
     final argument = arguments[i].trim();
-    final command = commands.firstWhere(
+    final command = ffCommands.firstWhere(
         (element) => element.contains(
               argument,
             ),
@@ -51,11 +52,16 @@ List<Command> initCommands(List<String> arguments) {
       if (command is Path) {
         i++;
         if (i >= arguments.length || arguments[i].startsWith('-')) {
-          throw Exception('miss value of path');
+          print(red.wrap('Miss value of [--path].'));
+          return null;
         }
         command.value = arguments[i];
       }
       result.add(command);
+    } else {
+      print(red.wrap(
+          'Could not find an option or flag "$argument". ${yellow.wrap('ff_route --help')}.'));
+      return null;
     }
   }
   return result;